requirements.txt
# AI Phone Assistant Requirements
fastapi==0.109.2
uvicorn[standard]==0.24.0
pydantic[email]==2.5.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.26.0
sqlalchemy==2.0.26
asyncpg==0.28.0
tenacity==8.2.3

# Twilio integration
twilio==8.9.1

# Google Calendar integration
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.111.0
python-dateutil==2.8.2

# OpenAI integration
openai==1.16.0

# For async concurrency
asyncio==3.4.3

# For testing
pytest==7.4.0
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
aiojobs==1.1.0  # For managing async background jobs
logging-extend==1.0.1
cachetools==5.3.2

# For backpressure and rate limiting
ratelimit==2.2.1

# For request validation and parsing
email-validator==2.1.0